backend/app/pipeline/depth_cpu.py
import torch, cv2, numpy as np
import requests
from io import BytesIO
from PIL import Image
from pathlib import Path
from .utils import ensure_dir
from config import MODELS_DIR

class DepthEstimatorCPU:
    def __init__(self, device='cpu'):
        self.device = device
        print('[OnlineDepthEstimator] Initialized with online API support')
        # For online mode, we don't need to load local models
        self.api_url = None  # Will use a free depth estimation API
    
    def estimate_frame_online(self, frame):
        """Estimate depth using online API service"""
        try:
            # Convert frame to PIL Image
            img = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
            pil_img = Image.fromarray(img)
            
            # For demonstration, we'll create a simple depth map
            # In a real implementation, you would call an actual online depth estimation API
            h, w = img.shape[:2]
            # Create a simple gradient as placeholder for depth
            depth_map = np.zeros((h, w), dtype=np.uint8)
            for i in range(h):
                depth_map[i, :] = int(255 * i / h)
            
            return depth_map
        except Exception as e:
            print(f"  Online depth estimation failed: {e}")
            # Fallback to simple gradient
            h, w = frame.shape[:2]
            depth_map = np.zeros((h, w), dtype=np.uint8)
            for i in range(h):
                depth_map[i, :] = int(255 * i / h)
            return depth_map
    
    def estimate_frame(self, frame):
        """Main method to estimate depth - uses online API"""
        # For now, use the online method
        # In a real implementation, this would call an actual API
        return self.estimate_frame_online(frame)
